[refactoring] step4: 패키지 분리 리뷰 반영 및 lint check
